+++ lib/features/theory/domain/analysis/chord_candidate_ranking.dart
@@ -965,6 +965,7 @@ class _CandidateFeatures {
       ChordToneRole.eleven => ChordExtension.eleven,
       ChordToneRole.thirteenth => ChordExtension.thirteen,
       ChordToneRole.add9 => ChordExtension.add9,
+      ChordToneRole.addSharp9 => ChordExtension.addSharp9,
       ChordToneRole.add11 => ChordExtension.add11,
       ChordToneRole.add13 => ChordExtension.add13,
       _ => null,